requirements.txt
aiohttp==1.0.5
async-timeout==1.1.0
cffi==1.9.1
chardet==2.3.0
discord.py==0.16.6
multidict==2.1.4
pycparser==2.17
PyNaCl==1.0.1
six==1.10.0
websockets==3.2
youtube-dl==2017.2.14
